# docs/dz8/swagger.yaml
openapi: 3.0.0
info:
  title: Микросервисная платформа электронной коммерции
  description: |
    API для микросервисной платформы электронной коммерции, включающей сервисы заказов, 
    оплаты, доставки, склада, биллинга и уведомлений
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Сервис заказов
  - url: http://localhost:8081
    description: Сервис биллинга
  - url: http://localhost:8082
    description: Сервис уведомлений
  - url: http://localhost:8083
    description: Сервис платежей
  - url: http://localhost:8084
    description: Сервис склада
  - url: http://localhost:8085
    description: Сервис доставки
...
paths:
...
  /api/v1/users:
    post:
      tags:
        - Orders
      summary: Создание пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                first_name:
                  type: string
                last_name:
                  type: string
              required:
                - username
                - email
      responses:
        '201':
          description: Пользователь создан
        '400':
          description: Ошибка в запросе
  /api/v1/orders:
    post:
      tags:
        - Orders
      summary: Создание заказа
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      quantity:
                        type: integer
                      price:
                        type: number
                        format: float
                delivery:
                  type: object
                  properties:
                    address:
                      type: string
                    time_slot_id:
                      type: string
                    zone_id:
                      type: string
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  status:
                    type: string
                  items:
                    type: array
                    items:
                      type: object
        '400':
          description: Ошибка в запросе
        '401':
          description: Пользователь не авторизован
  /api/v1/orders/{id}:
    get:
      tags:
        - Orders
      summary: Получение заказа по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  status:
                    type: string
                  items:
                    type: array
                    items:
                      type: object
        '404':
          description: Заказ не найден
  /api/v1/users/{id}/orders:
    get:
      tags:
        - Orders
      summary: Получение списка заказов пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
        '403':
          description: Доступ запрещен

  # Сервис биллинга
...
  /api/v1/payments/{id}:
    get:
      tags:
        - Payments
      summary: Получение платежа по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о платеже
        '404':
          description: Платеж не найден
  /api/v1/payments/process:
    post:
      tags:
        - Payments
      summary: Обработка платежа
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: integer
                user_id:
                  type: integer
                amount:
                  type: number
                  format: float
                payment_method:
                  type: string
      responses:
        '200':
          description: Платеж обработан
        '400':
          description: Ошибка в запросе
  /api/v1/payments/{id}/cancel:
    post:
      tags:
        - Payments
      summary: Отмена платежа
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Платеж отменен
        '400':
          description: Ошибка в запросе
  /api/v1/payments/by-order/{order_id}:
    get:
      tags:
        - Payments
      summary: Получение платежа по ID заказа
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о платеже
        '404':
          description: Платеж не найден
  /api/v1/payments/by-customer/{user_id}:
    get:
      tags:
        - Payments
      summary: Получение платежей пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список платежей
        '403':
          description: Доступ запрещен

  # Сервис платежей - Внутренние API
  /internal/payments/process:
    post:
      tags:
        - Payments
        - Internal
      summary: Внутренняя обработка платежа
      description: Эндпоинт для вызова другими сервисами, не требует стандартной аутентификации.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Платеж обработан
        '400':
          description: Ошибка в запросе
        '500':
          description: Внутренняя ошибка сервера
  /internal/payments/by-order/{order_id}:
    get:
      tags:
        - Payments
        - Internal
      summary: Внутреннее получение платежа по ID заказа
      description: Эндпоинт для вызова другими сервисами, не требует стандартной аутентификации.
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о платеже
        '404':
          description: Платеж не найден
        '500':
          description: Внутренняя ошибка сервера
  /internal/payments/{id}/cancel:
    post:
      tags:
        - Payments
        - Internal
      summary: Внутренняя отмена платежа
      description: Эндпоинт для вызова другими сервисами, не требует стандартной аутентификации.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Платеж отменен
        '400':
          description: Ошибка в запросе
        '500':
          description: Внутренняя ошибка сервера

  # Сервис склада
...
  /api/v1/warehouse/order/{order_id}:
    get:
      tags:
        - Warehouse
      summary: Получение резерваций для заказа
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список резерваций
        '404':
          description: Резервации не найдены

  # Сервис склада - Внутренние API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaymentRequest:
      type: object
      properties:
        order_id:
          type: integer
        user_id:
          type: integer
        amount:
          type: number
          format: float
        payment_method:
          type: string
...